+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- `components/FloatingButtonLayout` fixed issues with keyboard sometimes (and especially on some devices) not pushing/dragging the floating button with it
+
 ## [0.2.9 - 2020-06-05]
 
 ### Changed